server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.2.0",
    "bcrypt==4.0.1",
    "email-validator>=2.3.0",
    "fastapi>=0.122.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pymysql>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
    "pytest>=9.0.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]
